Add vendor logo support for the info-overview Distro's can use LOGO="vendorname" in their /etc/os-release file. If this is not specified then "budgie-logo" is assumed for the LOGO name prefix.
The chosen icon can be suffixed with "-flavor" for distros that are flavors of an overarching distro.
Two icons can optionally be specified - one for a dark panel; a dark panel logo should be suffixed with "-dark".
If no icon is found then the default budgie-desktop icon installed by
budgie-control-center is chosen.

Examples of a themed icon name:
mydistro
mydistro-dark
mydistro-flavor
mydistro-flavor-dark
budgie-logo-flavor
budgie-logo-flavor-dark
budgie-logo
budgie-logo-dark