metainfo: Update and modernize the AppStream data
This updates the AppStream metainfo file to use modern names and keys.
Part of this includes switching to RDNS names for the metainfo and
desktop files.

Note that we cannot fix the unknown desktop ID error because FreeDesktop
doesn't seem to think we exist. At least, we aren't on their page of
desktops.

Fixes https://github.com/BuddiesOfBudgie/budgie-control-center/issues/99

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>